// backend/src/enum/status.enum.ts
export enum TournamentStatus {
    DRAFT = 'DRAFT',
    ACTIVE = 'ACTIVE',
    COMPLETED = 'COMPLETED',
    CANCELLED = 'CANCELLED',
}

export enum MatchStatus {
    PENDING = 'PENDING',
    ONGOING = 'ONGOING',
    ENDED = 'ENDED',
    VALIDATED = 'VALIDATED',
}

export enum MatchesSessionStatus {
    OPEN = 'OPEN',
    CLOSED = 'CLOSED',
}